1.2.7 ForceOP Exploit added...
***New Discord Server for Updates and Friends:*** https://discord.gg/Az3G3p4ZCj
***ForceOPSign Module Added!***
- Added the ForceOPSign module which allows you to make signs in Creative mode that are pre-loaded with operator level commands. 
To use the ForceOPSign module, configure the options to make it execute the commands you want (default is /op yourname) then give the sign to an operator who is also in creative mode and have them place then right click on the sign.
The signs that ForceOPSign creates appear to be blank when placed, so the person may not know there is anything wrong with them.
- The commands that are executed by the ForceOPSign do not always appear in server chat response of operator commands. The operator will have no idea they just made you an operator when that command executes from the sign.
- There is an option to use the last three lines on the sign for a set of commands to auto grief a server. (These use /execute commands which only work in Minecraft 1.13 or higher)
- There is also an option to have the sign run a command to disable a server and have it crash over and over again. 

**OPERATORS CLICKING SIGNS CAN ONLY EXECUTE THE /OP COMMAND ON MINECRAFT VERSIONS GREATER THAN v1.12.2 and only if the server is PAPER**
(Paper versions grant atleast command permission level 3 on signs. Thats why we can Force OP on all Paper versions including the most recent.)

**IN VANILLA SERVERS AND PAPER VERSIONS LESS THAN OR EQUAL TO v1.12.2 THEN ANYONE CAN ONLY EXECUTE PERMISSION LEVEL 2 COMMANDS WHEN CLICKING SIGNS**
This is still really good and fun. You can run permission level two commands.
**Any Player** can click these signs after the OP places and it **executes** the commands. (Operators could have fun trolling their players with this!)
This includes /fill commands and even /setblock which could allow you to insert a malicious command block into the server in a spot that the admin does not know about.
Gamerule commands are also permission level 2 meaning that you could get an admin to change any gamerule you want without them knowing, just use "/gamerule sendCommandFeedback false" as a command before the rest.


Demonstration video here:
https://www.youtube.com/watch?v=YICXSxkvBfQ

More details and a standalone version of the mod is here:
https://github.com/etianl/ForceOPSign

I also made a build of Meteor Client for 1.20.4 from the latest source available for that Minecraft version. This is required for Trouser-Streak for 1.20.4!
https://github.com/etianl/meteor-client-1.20.4/releases/tag/0.5.6


**OTHER UPDATES**
- Added a Render Distance option to **Hole/Tunnel/StairsESP** which removes the rendered boxes that are outside of the render distance that you define. This fixes the lag due to the boxes staying loaded forever
- Added a Command Delay option to the **AutoCommand** module. This adds a delay between listed commands. 
- Removed the FULL AUTO option from **AutoCommand** and replaced it with an option called "Loop Commands" which decides whether or not to loop the defined commands.
- Fixed BoatKill not killing things for some reason.
- Updated the **.world command** to show your permission level for commands.