v1.2.9 ActivatedSpawnerDetector added, and bugfixes
**BaseFinder Changes/Fixes**
- Moved Hoppers from block list 1 to list 6 to prevent false positives with Trial Chambers. (Trouser for 1.21.3 only)
- Moved all beds from block lists to list 5 due to Trial Chambers now having beds of randomized colors in their chambers. (Trouser for 1.21.3 only)
- Fixed the Entity Cluster detector not working. It detects if there are large clusters of certain entities (such as dozens of chickens in a farm or something)

**New ActivatedSpawnerDetector Module!**
This new module can detect if a player was ever near a spawner block. 
It does so by checking the countdown value on the next spawn which gets changed from a certain value that it generates with.
The intended use for this module would be on anarchy servers where people try to hide their items within chests in dungeons and mineshafts.
This allows you to see if a dungeon or mineshaft has been visited even if the player who visited has not modified ANY of the natural blocks.