v1.3.2 PotESP module added for Decorated Pots
**PotESP Module Added!**
This module finds the dank pots for you, in Minecraft. By "dank pots" I actually mean Decorated Pots which contain items that are not natural. 
The module will tell you the location of the pot, and what is inside of it.
I was thinking some people might want to stash things in Trial Chamber pots thinking they are sneaky.