1.4.4 ElytraCount added, merged pr from crosby
- Added ElytraCount HUD option. This module tells you just how many Elytra you have left in your inventory. Thank you to https://github.com/Nataani3825 for adding this!
- Merged pull request https://github.com/etianl/Trouser-Streak/pull/89 from https://github.com/RacoonDog to make the build.gradle file better. Thanks! 
- Made the "Check Air Disturbances" option in **ActivatedSpawnerDetector** false by default because there are some false positives with it.
On a side note, I recently confirmed that the Book and Quill dupe only works on server versions 1.21.1 and less! 
If the server is 1.21.1 and is **NOT** vanilla then the dupe will not work. If the server is less than 1.21.1 or is 1.21.1 vanilla then the dupe will work!
Issue: https://github.com/etianl/Trouser-Streak/issues/77